Workaround: comparison of different form versions
wppsvr was panicking because the compare code assumes all messages of
the same type have parallel sets of fields.  But that isn't actually
true if they have different versions.

This commit works around the problem by adding a null field to older
versions of RACES-MAR to satisfy the bad expectation.